client-2/server/requirements.txt
fastapi
uvicorn[standard]
python-multipart           # file uploads
pydantic
openai                    # if using OpenAI LLMs
google-generativeai       # if using Gemini
anthropic                 # Claude SDK
ffmpeg-python             # if you’re slicing video
python-dotenv             # for .env config
aiofiles                  # async file handling
httpx                     # for making async API calls
tqdm                      # progress bars
nltk                      # for smarter chunking
pillow                    # optional, if working with thumbnails
moviepy                   # better video processing
numpy                     # for audio/video processing
soundfile